src/liboping.c: Added the `username' field which holds the name provided by the user.
Alex Brooks has reported that he cannot reliably remove hosts he added without
a fully qualified domain name, such as `foobar'. The reason is that the
`hostname' the user provided is replaced with the `canonical hostname', if they
differ. To fix this problem I introduced a new fields `username' which holds
the hostname as it was supplied by the user. This is the only name checked when
removing a host now.

Applications may access this name using the new `PING_INFO_USERNAME' option for
`ping_iterator_get_info'.