Added support for PING_OPT_DEVICE.
This option may be used to set the outgoing network device to be used. The
value passed must be a char-pointer to a null-terminated string specifying an
interface name (e.g. eth0).

Please note that this might not be supported by all operating systems. In that
case, ping_setopt() sets the error to "operation not supported".