Mac OX S 10.7 compile fixes.
I've attached a patch that's required to be able to compile liboping on
MacOS 10.7 (Lion).  The issue is that the IPv6 API's are changing from
RFC 2292 to 3542, and Apple's headers require you to state which API you
want to use by defining a constant.  I've changed it to use either based
on the presence of constants, but to prefer 3542 on MacOS.

One problem I see with this is that if you have both multicast and
unicast TTL's set, then this won't report both of them, but I'm not sure
if that's even valid and it would require more significant rework.
Also, I noticed you have fixed the IP_RECVTOS issue in head.  That patch
is also required to be able to compile on Solaris and MacOS.

Signed-off-by: Florian Forster <octo@verplant.org>