sound output clarification
I recently had to use audible ping again and ran through the procedure
from the manpage, but sound was still not coming out. I ended up
figuring out system sounds were muted, yet this wasn't in the
checklist.

This simply adds an extra step in the manpage to ensure the user
checks that part as well.