Attach /v4 or /v6 to visible hostnames.
With the "multiping" patch, you end up seeing a hostname multiple times
in the "noping" graphical display, and it's not easy to see right away
whether the IPv4 or IPv6 address has loss.  So, extend hostnames with
"/v4" or "/v6" to make this visible.

No option to turn this on or off has been implemented so far, but it
could be made dependant on "-4"/"-6" (no suffix needed) or on an extra
option.

Signed-off-by: Gert Doering <gert@greenie.muc.de>